Added textures Added glfw alternative to sdl
Adding glfw was in first order done to bring in opengl interop, but i gave up opengl interop as when i finally got it seemingly working, launching kernels crashed with a nonsensical error that i couldn't track down the cause of.
Keeping glfw because it wrecks sdl in terms of performance since i get hardware acceleration of drawing images to screen.